feat: 🎸 Split home page into components
